-get_animation works as it should now.

+++ modules/animation_utils/core_utils.cpp
@@ -46,9 +46,8 @@ void Summator::_bind_methods() {
 	ClassDB::bind_method(D_METHOD("add", "value"), &Summator::add);
 	ClassDB::bind_method(D_METHOD("reset"), &Summator::reset);
 	ClassDB::bind_method(D_METHOD("get_total"), &Summator::get_total);
-    ClassDB::bind_method(D_METHOD("get_animation"), &Summator::get_animation);
-    ClassDB::bind_method(D_METHOD("get_node"), &Summator::get_node);
-
+	ClassDB::bind_method(D_METHOD("get_animation"), &Summator::get_animation);
+	ClassDB::bind_method(D_METHOD("get_node"), &Summator::get_node);
 }
 
 Summator::Summator() {
@@ -59,39 +58,36 @@ Summator::Summator() {
  * @brief Summator::get_animation
  * @return
  */
-Ref<Animation> Summator::get_animation(NodePath node_path, String text, String delimeter)
-{
-    animation.instance();
-    int track_index = animation->add_track(Animation::TrackType::TYPE_VALUE, -1);
-    animation->track_set_path(track_index, node_path);
-    int current_transition = 0.0;
-    String current_text = "";
-    //TODO:When doing this from GDScript, it returns a PoolStringArray.
-    //PoolStringArray is supposed to be faster than Vector and I would like to use it, but don't know
-    //how at the moment.
-    Vector<String> tokens  = text.split(delimeter);
+Ref<Animation> Summator::get_animation(NodePath node_path, String text, String delimeter) {
+	animation.instance();
+	int track_index = animation->add_track(Animation::TrackType::TYPE_VALUE, -1);
+	animation->track_set_path(track_index, node_path);
+	float current_transition = 0.0;
+	String current_text = "";
+	//TODO:When doing this from GDScript, it returns a PoolStringArray.
+	//PoolStringArray is supposed to be faster than Vector and I would like to use it, but don't know
+	//how at the moment.
+	Vector<String> tokens = text.split(delimeter);
 
-    for(int i = 0; i<tokens.size();i++)
-    {
-        current_text = current_text + tokens[i] + delimeter;
+	for (int i = 0; i < tokens.size(); i++) {
+		current_text = current_text + tokens[i] + delimeter;
 
-        //This is very inefficient at the moment, but still learning the API...
-        Variant v{current_text};
+		//This is very inefficient at the moment, but still learning the API...
+		Variant v{ current_text };
 
-        animation->track_insert_key(track_index, current_transition, v);
-        current_transition = current_transition + 0.5;
-        animation->set_length(animation->get_length() + 0.5);
-    }
+		animation->track_insert_key(track_index, current_transition, v);
+		current_transition = current_transition + 0.5;
+		animation->set_length(animation->get_length() + 0.5);
+	}
 
-    return animation;
+	return animation;
 }
 
 /**
  * @brief Summator::get_node
  * @return
  */
-Node* Summator::get_node()
-{
-    node = std::make_unique<Node>();
-    return node.get();
+Node *Summator::get_node() {
+	node = std::make_unique<Node>();
+	return node.get();
 }
